+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/model/Markers.java
@@ -1,0 +1,11 @@
+package de.codescape.jira.plugins.multiplesubtasks.model;
+
+/**
+ * Markers for attributes with a special handling on subtask creation.
+ */
+public class Markers {
+
+    public static final String INHERIT_MARKER = "@inherit";
+    public static final String CURRENT_MARKER = "@current";
+
+}