+++ docs/changelog.md
@@ -17,6 +17,7 @@ Click on the version in the following list to see all changes since the previous
 * docs: improve documentation for custom fields
 * feat: extend syntax and allow to set fix version and affected version (#56)
 * feat: allow to use @inherit on fixVersion and affectedVersion (#72)
+* feat: extend syntax and allow to set due date (#53)
 
 ### [23.02.0] - 2023-02-05